extract Fieldset.tsx
