+++ src/pages/app.tsx
@@ -5,8 +5,6 @@ import Avatar from "@material-ui/core/Avatar"
 import Button from "@material-ui/core/Button"
 import ButtonGroup from "@material-ui/core/ButtonGroup"
 import Box from "@material-ui/core/Box"
-import Paper from "@material-ui/core/Paper"
-import { Divider } from "@material-ui/core"
 import { makeStyles } from "@material-ui/core/styles"
 import styled, { css, createGlobalStyle } from "styled-components"
